Compile boost for ARM Linux.
This patch introduces a new configuration setting called "linux_arm"
and updates the select() calls to enable compilation for ARM Linux.

It is an unfortunate limitation of bazel that select() does not work
in macros, and so the straightforward application of

    exclude_src = select(...)

is not possible.  For this reason compiling the "log" library requires
a workaround using cc_library() directly for the dump_ssse3 modules.